requirements.txt
cycler==0.10.0
defusedxml==0.4.1
diff-match-patch==20121119
dj-database-url==0.3.0
dj-static==0.0.6
Django==1.8
django-appconf==1.0.1
django-compressor==1.5
django-easy-pjax==1.2.0
django-flat-theme==0.9.3
django-image-cropping==1.0.1
django-import-export==0.4.5
easy-thumbnails==2.2
google-api-python-client==1.6.1
httplib2==0.9.2
numpy==1.12.0
oauth2client==4.0.0
Pillow==2.9.0
pyasn1==0.1.9
pyasn1-modules==0.0.8
PyJWT==1.3.0
pyparsing==2.1.10
python-dateutil==2.6.0
pytz==2016.10
requests==2.5.1
rsa==3.4.2
seaborn==0.7.1
six==1.9.0
static3==0.6.1
tablib==0.11.2
Unidecode==0.4.17
uritemplate==3.0.0
